LayoutTests/fast/forms/range/range-hit-test-with-padding-expected.txt

Clicking middle of input type=range with padding should set the value to middle.

PASS input.value is "500"
PASS successfullyParsed is true

TEST COMPLETE
